databene-commons-core/src/test/java/org/databene/commons/converter/Locale2StringConverterTest.java
/*
 * (c) Copyright 2010-2011 by Volker Bergmann. All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, is permitted under the terms of the
 * GNU General Public License (GPL).
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 * WITHOUT A WARRANTY OF ANY KIND. ALL EXPRESS OR IMPLIED CONDITIONS,
 * REPRESENTATIONS AND WARRANTIES, INCLUDING ANY IMPLIED WARRANTY OF
 * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE OR NON-INFRINGEMENT, ARE
 * HEREBY EXCLUDED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
 * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
 * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
 * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
 * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
 * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 * POSSIBILITY OF SUCH DAMAGE.
 */

package org.databene.commons.converter;

import static org.junit.Assert.assertEquals;



import java.util.Locale;

import org.databene.commons.ConversionException;
import org.junit.Test;

/**
 * Tests the {@link Locale2StringConverter}.<br/><br/>
 * Created: 25.02.2010 23:43:52
 * @since 0.5.0
 * @author Volker Bergmann
 */
public class Locale2StringConverterTest extends AbstractConverterTest {

	public Locale2StringConverterTest() {
	    super(Locale2StringConverter.class);
    }

	@Test
    public void test() throws ConversionException {
		Locale2StringConverter converter = new Locale2StringConverter();
        assertEquals("de", converter.convert(Locale.GERMAN));
        assertEquals("de_DE", converter.convert(Locale.GERMANY));
    }
	
}
